chore: 가독성을 위한 패키지 구조 변경
