--- app/src/main/java/com/towitty/bookreport/ui/SettingsActivity.kt
@@ -1,15 +1,0 @@
-package com.towitty.bookreport.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.towitty.bookreport.R
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class SettingsActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
-    }
-}
