--- app/src/main/java/com/towitty/bookreport/ui/navigation/Routes.kt
@@ -1,7 +1,0 @@
-package com.towitty.bookreport.ui.navigation
-
-object Routes {
-    const val BOOK_INFO_DETAIL: String = "book_info_detail"
-    const val BOOK_SEARCH_FOR_BOOK_REPORT: String = "book_report_book_search"
-    const val DIRECTLY_BOOK_REPORT = "directly_book_report"
-}
